+++ pmd/src/site/markdown/changelog.md
@@ -64,6 +64,7 @@
 * Fixed [bug 1154]: Call super onPause when there is no super
 * Fixed [bug 1155]: maven pmd plugin does not like empty rule sets
 * Fixed [bug 1159]: false positive UnusedFormalParameter readObject(ObjectInputStream) if not used
+* Fixed [bug 1164]: Violations are not suppressed with @java.lang.SuppressWarnings("all")
 
 [bug  881]: https://sourceforge.net/p/pmd/bugs/881
 [bug 1059]: https://sourceforge.net/p/pmd/bugs/1059
@@ -90,6 +91,7 @@
 [bug 1154]: https://sourceforge.net/p/pmd/bugs/1154
 [bug 1155]: https://sourceforge.net/p/pmd/bugs/1155
 [bug 1159]: https://sourceforge.net/p/pmd/bugs/1159
+[bug 1164]: https://sourceforge.net/p/pmd/bugs/1164